Fixed binarySearch() read desc.
binary search kept outputting 4 so I found out that i was returning values within the if statements.
I was also supposed to compare v{mid} to the elem not the end or start.
I also started the function with checking if v[mid] is equal to mid as that made much more sense in my head